Shifted test files to pycode
